# Remove dynamic dispatch in favor of switch