--- a/jmespath.js
+++ b/jmespath.js
@@ -1677,6 +1677,5 @@
   exports.tokenize = tokenize;
   exports.compile = compile;
   exports.search = search;
-  exports.Parser = Parser;
   exports.strictDeepEqual = strictDeepEqual;
 })(typeof exports === "undefined" ? this.jmespath = {} : exports);