--- a/package.json
+++ b/package.json
@@ -10,11 +10,11 @@
   "homepage": "https://github.com/jmespath/jmespath.js",
   "contributors": [],
   "devDependencies": {
-    "grunt": "^0.4.4",
+    "grunt": "^0.4.5",
     "grunt-contrib-jshint": "^0.11.0",
-    "mocha": "^2.1.0",
-    "grunt-contrib-uglify": "^0.7.0",
-    "grunt-eslint": "^6.0.0"
+    "grunt-contrib-uglify": "^0.11.1",
+    "grunt-eslint": "^17.3.1",
+    "mocha": "^2.1.0"
   },
   "dependencies": {},
   "main": "jmespath.js",