This PR pulls in a portion of the commits from (https://github.com/jmespath/jmespath.js/tree/reduce-size), which is work done to reduce the minified file size.

I'm not 100% sold on the commits from the `reduce-size` branch, so I've split parts of that branch into this PR, which represents commits I'm ok with merging to master.

The high level change is to switch from dynamic dispatch based on type names to a big switch statement.  This gives two benefits:

* Faster execution time
* Small minified (uncompressed) file size.

Benchmarks from master vs. this PR:

#### master branch

```
$ node perf.js
Mean time: 0.002411msec Parser#single_expr x 414,688 ops/sec ±1.62% (93 runs sampled)
Mean time: 0.005547msec Parser#single_subexpr x 180,273 ops/sec ±1.54% (93 runs sampled)
Mean time: 0.119469msec Parser#deeply_nested_50 x 8,370 ops/sec ±1.74% (91 runs sampled)
Mean time: 0.140929msec Parser#deeply_nested_50_index x 7,096 ops/sec ±1.79% (92 runs sampled)
Mean time: 0.008148msec Parser#basic_list_projection x 122,732 ops/sec ±1.94% (93 runs sampled)
```

#### perf-improve branch

```
$ node perf.js
Mean time: 0.002136msec Parser#single_expr x 468,057 ops/sec ±1.45% (95 runs sampled)
Mean time: 0.004363msec Parser#single_subexpr x 229,209 ops/sec ±1.97% (90 runs sampled)
Mean time: 0.080628msec Parser#deeply_nested_50 x 12,403 ops/sec ±1.52% (89 runs sampled)
Mean time: 0.076382msec Parser#deeply_nested_50_index x 13,092 ops/sec ±7.08% (88 runs sampled)
Mean time: 0.003640msec Parser#basic_list_projection x 274,699 ops/sec ±1.58% (96 runs sampled)
```

Minified file sizes:

* master: `23179` bytes
* perf-improve: `22037` bytes